requirements.txt
Django==3.2.23
paho-mqtt==1.6.1
djangorestframework==3.14.0
django-cors-headers==3.14.0
PyJWT==2.8.0
Pillow==10.1.0